refactor pickers to own components
