+++ .babelrc
@@ -7,8 +7,7 @@
   "comments": false,
   "env": {
     "test": {
-      "presets": ["env", "stage-2"],
-      "plugins": [ "istanbul" ]
+      "presets": ["env", "stage-2"]
     }
   }
 }
