+++ package.json
@@ -28,9 +28,7 @@
     "lint": "eslint --ext .js,.vue src test/unit/specs",
     "prepublish": "npm run build"
   },
-  "dependencies": {
-    "vue-server-renderer": "^2.5.16"
-  },
+  "dependencies": {},
   "devDependencies": {
     "@vue/test-utils": "^1.0.0-beta.12",
     "autoprefixer": "^6.7.2",
@@ -38,7 +36,6 @@
     "babel-eslint": "^7.1.1",
     "babel-jest": "^22.4.3",
     "babel-loader": "^6.2.10",
-    "babel-plugin-istanbul": "^4.1.1",
     "babel-plugin-transform-runtime": "^6.22.0",
     "babel-preset-env": "^1.3.2",
     "babel-preset-stage-2": "^6.22.0",
@@ -82,6 +79,7 @@
     "vue": "^2.3.3",
     "vue-jest": "^2.2.1",
     "vue-loader": "^12.1.0",
+    "vue-server-renderer": "^2.5.16",
     "vue-style-loader": "^3.0.1",
     "vue-template-compiler": "^2.3.3",
     "webpack": "^2.6.1",
