+++ src/components/Datepicker.vue
@@ -3,86 +3,71 @@
     <date-input :props="inputProps" @showCalendar="showCalendar" />
 
     <!-- Day View -->
-    <template v-if="allowedToShowView('day')">
-      <div :class="[calendarClass, 'vdp-datepicker__calendar']" v-show="showDayView" v-bind:style="calendarStyle">
-          <header>
-              <span
-                  @click="isRtl ? nextMonth() : previousMonth()"
-                  class="prev"
-                  v-bind:class="{ 'disabled' : isRtl ? nextMonthDisabled(pageTimestamp) : previousMonthDisabled(pageTimestamp) }">&lt;</span>
-              <span @click="showMonthCalendar" :class="allowedToShowView('month') ? 'up' : ''">{{ isYmd ? currYear : currMonthName }} {{ isYmd ? currMonthName : currYear }}</span>
-              <span
-                  @click="isRtl ? previousMonth() : nextMonth()"
-                  class="next"
-                  v-bind:class="{ 'disabled' : isRtl ? previousMonthDisabled(pageTimestamp) : nextMonthDisabled(pageTimestamp) }">&gt;</span>
-          </header>
-          <div :class="isRtl ? 'flex-rtl' : ''">
-            <span class="cell day-header" v-for="d in daysOfWeek" :key="d.timestamp">{{ d }}</span>
-            <template v-if="blankDays > 0">
-              <span class="cell day blank" v-for="d in blankDays" :key="d.timestamp"></span>
-            </template><!--
-            --><span class="cell day"
-                v-for="day in days"
-                :key="day.timestamp"
-                track-by="timestamp"
-                v-bind:class="dayClasses(day)"
-                @click="selectDate(day)">{{ day.date }}</span>
-          </div>
-      </div>
-    </template>
+    <picker-day
+      v-if="allowedToShowView('day')"
+      :pageDate="pageDate"
+      :selectedDate="selectedDate"
+      :showDayView="showDayView"
+      :fullMonthName="fullMonthName"
+      :allowedToShowView="allowedToShowView"
+      :disabled="disabled"
+      :highlighted="highlighted"
+      :calendarClass="calendarClass"
+      :calendarStyle="calendarStyle"
+      :translation="translation"
+      :pageTimestamp="pageTimestamp"
+      :isRtl="isRtl"
+      :mondayFirst="mondayFirst"
+      @changedMonth="setPageDate"
+      @selectDate="selectDate"
+      @showMonthCalendar="showMonthCalendar"
+      @selectedDisabled="$emit('selectedDisabled')"
+    />
 
     <!-- Month View -->
-    <template v-if="allowedToShowView('month')">
-      <div :class="[calendarClass, 'vdp-datepicker__calendar']" v-show="showMonthView" v-bind:style="calendarStyle">
-          <header>
-              <span
-                  @click="previousYear"
-                  class="prev"
-                  v-bind:class="{ 'disabled' : previousYearDisabled(pageTimestamp) }">&lt;</span>
-              <span @click="showYearCalendar" :class="allowedToShowView('year') ? 'up' : ''">{{ getPageYear() }}</span>
-              <span
-                  @click="nextYear"
-                  class="next"
-                  v-bind:class="{ 'disabled' : nextYearDisabled(pageTimestamp) }">&gt;</span>
-          </header>
-          <span class="cell month"
-              v-for="month in months"
-              :key="month.timestamp"
-              track-by="timestamp"
-              v-bind:class="{ 'selected': month.isSelected, 'disabled': month.isDisabled }"
-              @click.stop="selectMonth(month)">{{ month.month }}</span>
-      </div>
-    </template>
+    <picker-month
+      v-if="allowedToShowView('month')"
+      :pageDate="pageDate"
+      :selectedDate="selectedDate"
+      :showMonthView="showMonthView"
+      :allowedToShowView="allowedToShowView"
+      :disabled="disabled"
+      :calendarClass="calendarClass"
+      :calendarStyle="calendarStyle"
+      :translation="translation"
+      @selectMonth="selectMonth"
+      @showYearCalendar="showYearCalendar"
+      @changedYear="setPageDate"
+    />
 
     <!-- Year View -->
-    <template v-if="allowedToShowView('year')">
-      <div :class="[calendarClass, 'vdp-datepicker__calendar']" v-show="showYearView" v-bind:style="calendarStyle">
-          <header>
-              <span @click="previousDecade" class="prev"
-                  v-bind:class="{ 'disabled' : previousDecadeDisabled(pageTimestamp) }">&lt;</span>
-              <span>{{ getPageDecade() }}</span>
-              <span @click="nextDecade" class="next"
-                  v-bind:class="{ 'disabled' : nextMonthDisabled(pageTimestamp) }">&gt;</span>
-          </header>
-          <span
-              class="cell year"
-              v-for="year in years"
-              :key="year.timestamp"
-              track-by="timestamp"
-              v-bind:class="{ 'selected': year.isSelected, 'disabled': year.isDisabled }"
-              @click.stop="selectYear(year)">{{ year.year }}</span>
-      </div>
-    </template>
+    <picker-year
+      v-if="allowedToShowView('year')"
+      :pageDate="pageDate"
+      :selectedDate="selectedDate"
+      :showYearView="showYearView"
+      :allowedToShowView="allowedToShowView"
+      :disabled="disabled"
+      :calendarClass="calendarClass"
+      :calendarStyle="calendarStyle"
+      :translation="translation"
+      @selectYear="selectYear"
+      @changedDecade="setPageDate"
+    />
   </div>
 </template>
-
 <script>
-import DateUtils from '@/utils/DateUtils.js'
 import DateLanguages from '@/utils/DateLanguages.js'
 import DateInput from '@/components/DateInput'
+import PickerDay from '@/components/PickerDay'
+import PickerMonth from '@/components/PickerMonth'
+import PickerYear from '@/components/PickerYear'
 export default {
   components: {
-    DateInput
+    DateInput,
+    PickerDay,
+    PickerMonth,
+    PickerYear
   },
   props: {
     value: {
@@ -208,90 +193,7 @@ export default {
     translation () {
       return DateLanguages.translations[this.language]
     },
-    currMonthName () {
-      const monthName = this.fullMonthName ? this.translation.months.original : this.translation.months.abbr
-      return DateUtils.getMonthNameAbbr(this.pageDate.getMonth(), monthName)
-    },
-    currYear () {
-      return this.pageDate.getFullYear()
-    },
-    /**
-     * Returns the day number of the week less one for the first of the current month
-     * Used to show amount of empty cells before the first in the day calendar layout
-     * @return {Number}
-     */
-    blankDays () {
-      const d = this.pageDate
-      let dObj = new Date(d.getFullYear(), d.getMonth(), 1, d.getHours(), d.getMinutes())
-      if (this.mondayFirst) {
-        return dObj.getDay() > 0 ? dObj.getDay() - 1 : 6
-      }
-      return dObj.getDay()
-    },
-    daysOfWeek () {
-      if (this.mondayFirst) {
-        const tempDays = this.translation.days.slice()
-        tempDays.push(tempDays.shift())
-        return tempDays
-      }
-      return this.translation.days
-    },
-    days () {
-      const d = this.pageDate
-      let days = []
-      // set up a new date object to the beginning of the current 'page'
-      let dObj = new Date(d.getFullYear(), d.getMonth(), 1, d.getHours(), d.getMinutes())
-      let daysInMonth = DateUtils.daysInMonth(dObj.getFullYear(), dObj.getMonth())
-      for (let i = 0; i < daysInMonth; i++) {
-        days.push({
-          date: dObj.getDate(),
-          timestamp: dObj.getTime(),
-          isSelected: this.isSelectedDate(dObj),
-          isDisabled: this.isDisabledDate(dObj),
-          isHighlighted: this.isHighlightedDate(dObj),
-          isHighlightStart: this.isHighlightStart(dObj),
-          isHighlightEnd: this.isHighlightEnd(dObj),
-          isToday: dObj.toDateString() === (new Date()).toDateString(),
-          isWeekend: dObj.getDay() === 0 || dObj.getDay() === 6,
-          isSaturday: dObj.getDay() === 6,
-          isSunday: dObj.getDay() === 0
-        })
-        dObj.setDate(dObj.getDate() + 1)
-      }
-      return days
-    },
-    months () {
-      const d = this.pageDate
-      let months = []
-      // set up a new date object to the beginning of the current 'page'
-      let dObj = new Date(d.getFullYear(), 0, d.getDate(), d.getHours(), d.getMinutes())
-      for (let i = 0; i < 12; i++) {
-        months.push({
-          month: DateUtils.getMonthName(i, this.translation.months.original),
-          timestamp: dObj.getTime(),
-          isSelected: this.isSelectedMonth(dObj),
-          isDisabled: this.isDisabledMonth(dObj)
-        })
-        dObj.setMonth(dObj.getMonth() + 1)
-      }
-      return months
-    },
-    years () {
-      const d = this.pageDate
-      let years = []
-      // set up a new date object to the beginning of the current 'page'
-      let dObj = new Date(Math.floor(d.getFullYear() / 10) * 10, d.getMonth(), d.getDate(), d.getHours(), d.getMinutes())
-      for (let i = 0; i < 10; i++) {
-        years.push({
-          year: dObj.getFullYear(),
-          timestamp: dObj.getTime(),
-          isSelected: this.isSelectedYear(dObj),
-          isDisabled: this.isDisabledYear(dObj)
-        })
-        dObj.setFullYear(dObj.getFullYear() + 1)
-      }
-      return years
-    },
+
     calendarStyle () {
       return {
         position: this.isInline ? 'static' : undefined
@@ -305,9 +207,6 @@ export default {
     },
     isRtl () {
       return this.translation.rtl === true
-    },
-    isYmd () {
-      return this.translation.ymd === true
     }
   },
   methods: {
@@ -372,22 +271,25 @@ export default {
       return viewIndex >= minimumViewIndex && viewIndex <= maximumViewIndex
     },
     showDayCalendar () {
-      if (!this.allowedToShowView('day')) return false
-
+      if (!this.allowedToShowView('day')) {
+        return false
+      }
       this.close()
       this.showDayView = true
       this.addOutsideClickListener()
     },
     showMonthCalendar () {
-      if (!this.allowedToShowView('month')) return false
-
+      if (!this.allowedToShowView('month')) {
+        return false
+      }
       this.close()
       this.showMonthView = true
       this.addOutsideClickListener()
     },
     showYearCalendar () {
-      if (!this.allowedToShowView('year')) return false
-
+      if (!this.allowedToShowView('year')) {
+        return false
+      }
       this.close()
       this.showYearView = true
       this.addOutsideClickListener()
@@ -416,10 +318,6 @@ export default {
      * @param {Object} day
      */
     selectDate (day) {
-      if (day.isDisabled) {
-        this.$emit('selectedDisabled', day)
-        return false
-      }
       this.setDate(day.timestamp)
       if (!this.isInline) {
         this.close(true)
@@ -429,10 +327,6 @@ export default {
      * @param {Object} month
      */
     selectMonth (month) {
-      if (month.isDisabled) {
-        return false
-      }
-
       const date = new Date(month.timestamp)
       if (this.allowedToShowView('day')) {
         this.setPageDate(date)
@@ -449,10 +343,6 @@ export default {
      * @param {Object} year
      */
     selectYear (year) {
-      if (year.isDisabled) {
-        return false
-      }
-
       const date = new Date(year.timestamp)
       if (this.allowedToShowView('month')) {
         this.setPageDate(date)
@@ -464,319 +354,6 @@ export default {
           this.close(true)
         }
       }
-    },
-    /**
-     * @return {Number}
-     */
-    getPageDate () {
-      return this.pageDate.getDate()
-    },
-    /**
-     * @return {Number}
-     */
-    getPageMonth () {
-      return this.pageDate.getMonth()
-    },
-    /**
-     * @return {Number}
-     */
-    getPageYear () {
-      return this.pageDate.getFullYear()
-    },
-    /**
-     * @return {String}
-     */
-    getPageDecade () {
-      const decadeStart = Math.floor(this.pageDate.getFullYear() / 10) * 10
-      const decadeEnd = decadeStart + 9
-      return `${decadeStart} - ${decadeEnd}`
-    },
-    changeMonth (incrementBy) {
-      let date = this.pageDate
-      date.setMonth(date.getMonth() + incrementBy)
-      this.setPageDate(date)
-      this.$emit('changedMonth', date)
-    },
-    previousMonth () {
-      if (!this.previousMonthDisabled()) {
-        this.changeMonth(-1)
-      }
-    },
-    previousMonthDisabled () {
-      if (!this.disabled || !this.disabled.to) {
-        return false
-      }
-      let d = this.pageDate
-      return this.disabled.to.getMonth() >= d.getMonth() &&
-        this.disabled.to.getFullYear() >= d.getFullYear()
-    },
-    nextMonth () {
-      if (!this.nextMonthDisabled()) {
-        this.changeMonth(+1)
-      }
-    },
-    nextMonthDisabled () {
-      if (!this.disabled || !this.disabled.from) {
-        return false
-      }
-      let d = this.pageDate
-      return this.disabled.from.getMonth() <= d.getMonth() &&
-        this.disabled.from.getFullYear() <= d.getFullYear()
-    },
-    changeYear (incrementBy, emit = 'changedYear') {
-      let date = this.pageDate
-      date.setYear(date.getFullYear() + incrementBy)
-      this.setPageDate(date)
-      this.$emit(emit, date)
-    },
-    previousYear () {
-      if (!this.previousYearDisabled()) {
-        this.changeYear(-1)
-      }
-    },
-    previousYearDisabled () {
-      if (!this.disabled || !this.disabled.to) {
-        return false
-      }
-      return this.disabled.to.getFullYear() >= this.pageDate.getFullYear()
-    },
-    nextYear () {
-      if (!this.nextYearDisabled()) {
-        this.changeYear(1)
-      }
-    },
-    nextYearDisabled () {
-      if (!this.disabled || !this.disabled.from) {
-        return false
-      }
-      return this.disabled.from.getFullYear() <= this.pageDate.getFullYear()
-    },
-    previousDecade () {
-      if (!this.previousDecadeDisabled()) {
-        this.changeYear(-10, 'changeDecade')
-      }
-    },
-    previousDecadeDisabled () {
-      if (!this.disabled || !this.disabled.to) {
-        return false
-      }
-      return Math.floor(this.disabled.to.getFullYear() / 10) * 10 >= Math.floor(this.pageDate.getFullYear() / 10) * 10
-    },
-    nextDecade () {
-      if (!this.nextDecadeDisabled()) {
-        this.changeYear(10, 'changeDecade')
-      }
-    },
-    nextDecadeDisabled () {
-      if (!this.disabled || !this.disabled.from) {
-        return false
-      }
-      return Math.ceil(this.disabled.from.getFullYear() / 10) * 10 <= Math.ceil(this.pageDate.getFullYear() / 10) * 10
-    },
-    /**
-     * Whether a day is selected
-     * @param {Date}
-     * @return {Boolean}
-     */
-    isSelectedDate (dObj) {
-      return this.selectedDate && this.selectedDate.toDateString() === dObj.toDateString()
-    },
-    /**
-     * Whether a day is disabled
-     * @param {Date}
-     * @return {Boolean}
-     */
-    isDisabledDate (date) {
-      let disabled = false
-
-      if (typeof this.disabled === 'undefined') {
-        return false
-      }
-
-      if (typeof this.disabled.dates !== 'undefined') {
-        this.disabled.dates.forEach((d) => {
-          if (date.toDateString() === d.toDateString()) {
-            disabled = true
-            return true
-          }
-        })
-      }
-      if (typeof this.disabled.to !== 'undefined' && this.disabled.to && date < this.disabled.to) {
-        disabled = true
-      }
-      if (typeof this.disabled.from !== 'undefined' && this.disabled.from && date > this.disabled.from) {
-        disabled = true
-      }
-      if (typeof this.disabled.ranges !== 'undefined') {
-        this.disabled.ranges.forEach((range) => {
-          if (typeof range.from !== 'undefined' && range.from && typeof range.to !== 'undefined' && range.to) {
-            if (date < range.to && date > range.from) {
-              disabled = true
-              return true
-            }
-          }
-        })
-      }
-      if (typeof this.disabled.days !== 'undefined' && this.disabled.days.indexOf(date.getDay()) !== -1) {
-        disabled = true
-      }
-      if (typeof this.disabled.daysOfMonth !== 'undefined' && this.disabled.daysOfMonth.indexOf(date.getDate()) !== -1) {
-        disabled = true
-      }
-      if (typeof this.disabled.customPredictor === 'function' && this.disabled.customPredictor(date)) {
-        disabled = true
-      }
-      return disabled
-    },
-    /**
-     * Whether a day is highlighted (only if it is not disabled already except when highlighted.includeDisabled is true)
-     * @param {Date}
-     * @return {Boolean}
-     */
-    isHighlightedDate (date) {
-      if (!(this.highlighted && this.highlighted.includeDisabled) && this.isDisabledDate(date)) {
-        return false
-      }
-
-      let highlighted = false
-
-      if (typeof this.highlighted === 'undefined') {
-        return false
-      }
-
-      if (typeof this.highlighted.dates !== 'undefined') {
-        this.highlighted.dates.forEach((d) => {
-          if (date.toDateString() === d.toDateString()) {
-            highlighted = true
-            return true
-          }
-        })
-      }
-
-      if (this.isDefined(this.highlighted.from) && this.isDefined(this.highlighted.to)) {
-        highlighted = date >= this.highlighted.from && date <= this.highlighted.to
-      }
-
-      if (typeof this.highlighted.days !== 'undefined' && this.highlighted.days.indexOf(date.getDay()) !== -1) {
-        highlighted = true
-      }
-
-      if (typeof this.highlighted.daysOfMonth !== 'undefined' && this.highlighted.daysOfMonth.indexOf(date.getDate()) !== -1) {
-        highlighted = true
-      }
-
-      if (typeof this.highlighted.customPredictor === 'function' && this.highlighted.customPredictor(date)) {
-        highlighted = true
-      }
-
-      return highlighted
-    },
-    /**
-     * Whether a day is highlighted and it is the first date
-     * in the highlighted range of dates
-     * @param {Date}
-     * @return {Boolean}
-     */
-    isHighlightStart (date) {
-      return this.isHighlightedDate(date) &&
-        (this.highlighted.from instanceof Date) &&
-        (this.highlighted.from.getFullYear() === date.getFullYear()) &&
-        (this.highlighted.from.getMonth() === date.getMonth()) &&
-        (this.highlighted.from.getDate() === date.getDate())
-    },
-    /**
-     * Whether a day is highlighted and it is the first date
-     * in the highlighted range of dates
-     * @param {Date}
-     * @return {Boolean}
-     */
-    isHighlightEnd (date) {
-      return this.isHighlightedDate(date) &&
-        (this.highlighted.to instanceof Date) &&
-        (this.highlighted.to.getFullYear() === date.getFullYear()) &&
-        (this.highlighted.to.getMonth() === date.getMonth()) &&
-        (this.highlighted.to.getDate() === date.getDate())
-    },
-    /**
-     * Helper
-     * @param  {mixed}  prop
-     * @return {Boolean}
-     */
-    isDefined (prop) {
-      return typeof prop !== 'undefined' && prop
-    },
-    /**
-     * Whether the selected date is in this month
-     * @param {Date}
-     * @return {Boolean}
-     */
-    isSelectedMonth (date) {
-      return (this.selectedDate &&
-        this.selectedDate.getFullYear() === date.getFullYear() &&
-        this.selectedDate.getMonth() === date.getMonth())
-    },
-    /**
-     * Whether a month is disabled
-     * @param {Date}
-     * @return {Boolean}
-     */
-    isDisabledMonth (date) {
-      let disabled = false
-
-      if (typeof this.disabled === 'undefined') {
-        return false
-      }
-
-      if (typeof this.disabled.to !== 'undefined' && this.disabled.to) {
-        if (
-          (date.getMonth() < this.disabled.to.getMonth() && date.getFullYear() <= this.disabled.to.getFullYear()) ||
-          date.getFullYear() < this.disabled.to.getFullYear()
-        ) {
-          disabled = true
-        }
-      }
-      if (typeof this.disabled.from !== 'undefined' && this.disabled.from) {
-        if (
-          this.disabled.from &&
-          (date.getMonth() > this.disabled.from.getMonth() && date.getFullYear() >= this.disabled.from.getFullYear()) ||
-          date.getFullYear() > this.disabled.from.getFullYear()
-        ) {
-          disabled = true
-        }
-      }
-      return disabled
-    },
-    /**
-     * Whether the selected date is in this year
-     * @param {Date}
-     * @return {Boolean}
-     */
-    isSelectedYear (date) {
-      return this.selectedDate && this.selectedDate.getFullYear() === date.getFullYear()
-    },
-    /**
-     * Whether a year is disabled
-     * @param {Date}
-     * @return {Boolean}
-     */
-    isDisabledYear (date) {
-      let disabled = false
-      if (typeof this.disabled === 'undefined' || !this.disabled) {
-        return false
-      }
-
-      if (typeof this.disabled.to !== 'undefined' && this.disabled.to) {
-        if (date.getFullYear() < this.disabled.to.getFullYear()) {
-          disabled = true
-        }
-      }
-      if (typeof this.disabled.from !== 'undefined' && this.disabled.from) {
-        if (date.getFullYear() > this.disabled.from.getFullYear()) {
-          disabled = true
-        }
-      }
-
-      return disabled
     },
     /**
      * Set the datepicker value
@@ -795,6 +372,9 @@ export default {
       this.selectedDate = date
       this.setPageDate(date)
     },
+    /**
+     * Sets the date that the calendar should open on
+     */
     setPageDate (date) {
       if (!date) {
         if (this.openDate) {
@@ -819,19 +399,7 @@ export default {
         document.removeEventListener('click', this.clickOutside, false)
       }
     },
-    dayClasses (day) {
-      return {
-        'selected': day.isSelected,
-        'disabled': day.isDisabled,
-        'highlighted': day.isHighlighted,
-        'today': day.isToday,
-        'weekend': day.isWeekend,
-        'sat': day.isSaturday,
-        'sun': day.isSunday,
-        'highlight-start': day.isHighlightStart,
-        'highlight-end': day.isHighlightEnd
-      }
-    },
+
     init () {
       if (this.value) {
         this.setValue(this.value)
