+++ src/components/PickerDay.vue
@@ -1,0 +1,314 @@
+<template>
+  <div :class="[calendarClass, 'vdp-datepicker__calendar']" v-show="showDayView" :style="calendarStyle">
+    <header>
+      <span
+        @click="isRtl ? nextMonth() : previousMonth()"
+        class="prev"
+        :class="{ 'disabled' : isRtl ? nextMonthDisabled(pageTimestamp) : previousMonthDisabled(pageTimestamp) }">&lt;</span>
+      <span @click="showMonthCalendar" :class="allowedToShowView('month') ? 'up' : ''">{{ isYmd ? currYear : currMonthName }} {{ isYmd ? currMonthName : currYear }}</span>
+      <span
+        @click="isRtl ? previousMonth() : nextMonth()"
+        class="next"
+        :class="{ 'disabled' : isRtl ? previousMonthDisabled(pageTimestamp) : nextMonthDisabled(pageTimestamp) }">&gt;</span>
+    </header>
+    <div :class="isRtl ? 'flex-rtl' : ''">
+      <span class="cell day-header" v-for="d in daysOfWeek" :key="d.timestamp">{{ d }}</span>
+      <template v-if="blankDays > 0">
+        <span class="cell day blank" v-for="d in blankDays" :key="d.timestamp"></span>
+      </template><!--
+      --><span class="cell day"
+          v-for="day in days"
+          :key="day.timestamp"
+          :class="dayClasses(day)"
+          @click="selectDate(day)">{{ day.date }}</span>
+    </div>
+  </div>
+</template>
+<script>
+import DateUtils from '@/utils/DateUtils'
+export default {
+  props: {
+    showDayView: Boolean,
+    selectedDate: Date,
+    pageDate: Date,
+    pageTimestamp: Number,
+    fullMonthName: Boolean,
+    allowedToShowView: Function,
+    disabled: Object,
+    highlighted: Object,
+    calendarClass: String,
+    calendarStyle: Object,
+    translation: Object,
+    isRtl: Boolean,
+    mondayFirst: Boolean
+  },
+  computed: {
+    /**
+     * Returns an array of day names
+     * @return {String[]}
+     */
+    daysOfWeek () {
+      if (this.mondayFirst) {
+        const tempDays = this.translation.days.slice()
+        tempDays.push(tempDays.shift())
+        return tempDays
+      }
+      return this.translation.days
+    },
+    /**
+     * Returns the day number of the week less one for the first of the current month
+     * Used to show amount of empty cells before the first in the day calendar layout
+     * @return {Number}
+     */
+    blankDays () {
+      const d = this.pageDate
+      let dObj = new Date(d.getFullYear(), d.getMonth(), 1, d.getHours(), d.getMinutes())
+      if (this.mondayFirst) {
+        return dObj.getDay() > 0 ? dObj.getDay() - 1 : 6
+      }
+      return dObj.getDay()
+    },
+    /**
+     * @return {Object[]}
+     */
+    days () {
+      const d = this.pageDate
+      let days = []
+      // set up a new date object to the beginning of the current 'page'
+      let dObj = new Date(d.getFullYear(), d.getMonth(), 1, d.getHours(), d.getMinutes())
+      let daysInMonth = DateUtils.daysInMonth(dObj.getFullYear(), dObj.getMonth())
+      for (let i = 0; i < daysInMonth; i++) {
+        days.push({
+          date: dObj.getDate(),
+          timestamp: dObj.getTime(),
+          isSelected: this.isSelectedDate(dObj),
+          isDisabled: this.isDisabledDate(dObj),
+          isHighlighted: this.isHighlightedDate(dObj),
+          isHighlightStart: this.isHighlightStart(dObj),
+          isHighlightEnd: this.isHighlightEnd(dObj),
+          isToday: dObj.toDateString() === (new Date()).toDateString(),
+          isWeekend: dObj.getDay() === 0 || dObj.getDay() === 6,
+          isSaturday: dObj.getDay() === 6,
+          isSunday: dObj.getDay() === 0
+        })
+        dObj.setDate(dObj.getDate() + 1)
+      }
+      return days
+    },
+
+    currMonthName () {
+      const monthName = this.fullMonthName ? this.translation.months.original : this.translation.months.abbr
+      return DateUtils.getMonthNameAbbr(this.pageDate.getMonth(), monthName)
+    },
+
+    currYear () {
+      return this.pageDate.getFullYear()
+    },
+
+    isYmd () {
+      return this.translation.ymd && this.translation.ymd === true
+    }
+  },
+  methods: {
+    selectDate (date) {
+      if (date.isDisabled) {
+        this.$emit('selectedDisabled', date)
+        return false
+      }
+      this.$emit('selectDate', date)
+    },
+    /**
+     * @return {Number}
+     */
+    getPageMonth () {
+      return this.pageDate.getMonth()
+    },
+    /**
+     * @return {Number}
+     */
+    getPageYear () {
+      return this.pageDate.getYear()
+    },
+    showMonthCalendar () {
+      this.$emit('showMonthCalendar')
+    },
+    changeMonth (incrementBy) {
+      let date = this.pageDate
+      date.setMonth(date.getMonth() + incrementBy)
+      this.$emit('changedMonth', date)
+    },
+
+    previousMonth () {
+      if (!this.previousMonthDisabled()) {
+        this.changeMonth(-1)
+      }
+    },
+    previousMonthDisabled () {
+      if (!this.disabled || !this.disabled.to) {
+        return false
+      }
+      let d = this.pageDate
+      return this.disabled.to.getMonth() >= d.getMonth() &&
+        this.disabled.to.getFullYear() >= d.getFullYear()
+    },
+    nextMonth () {
+      if (!this.nextMonthDisabled()) {
+        this.changeMonth(+1)
+      }
+    },
+    nextMonthDisabled () {
+      if (!this.disabled || !this.disabled.from) {
+        return false
+      }
+      let d = this.pageDate
+      return this.disabled.from.getMonth() <= d.getMonth() &&
+        this.disabled.from.getFullYear() <= d.getFullYear()
+    },
+    /**
+     * Whether a day is selected
+     * @param {Date}
+     * @return {Boolean}
+     */
+    isSelectedDate (dObj) {
+      return this.selectedDate && this.selectedDate.toDateString() === dObj.toDateString()
+    },
+    /**
+     * Whether a day is disabled
+     * @param {Date}
+     * @return {Boolean}
+     */
+    isDisabledDate (date) {
+      let disabled = false
+
+      if (typeof this.disabled === 'undefined') {
+        return false
+      }
+
+      if (typeof this.disabled.dates !== 'undefined') {
+        this.disabled.dates.forEach((d) => {
+          if (date.toDateString() === d.toDateString()) {
+            disabled = true
+            return true
+          }
+        })
+      }
+      if (typeof this.disabled.to !== 'undefined' && this.disabled.to && date < this.disabled.to) {
+        disabled = true
+      }
+      if (typeof this.disabled.from !== 'undefined' && this.disabled.from && date > this.disabled.from) {
+        disabled = true
+      }
+      if (typeof this.disabled.ranges !== 'undefined') {
+        this.disabled.ranges.forEach((range) => {
+          if (typeof range.from !== 'undefined' && range.from && typeof range.to !== 'undefined' && range.to) {
+            if (date < range.to && date > range.from) {
+              disabled = true
+              return true
+            }
+          }
+        })
+      }
+      if (typeof this.disabled.days !== 'undefined' && this.disabled.days.indexOf(date.getDay()) !== -1) {
+        disabled = true
+      }
+      if (typeof this.disabled.daysOfMonth !== 'undefined' && this.disabled.daysOfMonth.indexOf(date.getDate()) !== -1) {
+        disabled = true
+      }
+      if (typeof this.disabled.customPredictor === 'function' && this.disabled.customPredictor(date)) {
+        disabled = true
+      }
+      return disabled
+    },
+    /**
+     * Whether a day is highlighted (only if it is not disabled already except when highlighted.includeDisabled is true)
+     * @param {Date}
+     * @return {Boolean}
+     */
+    isHighlightedDate (date) {
+      if (!(this.highlighted && this.highlighted.includeDisabled) && this.isDisabledDate(date)) {
+        return false
+      }
+
+      let highlighted = false
+
+      if (typeof this.highlighted === 'undefined') {
+        return false
+      }
+
+      if (typeof this.highlighted.dates !== 'undefined') {
+        this.highlighted.dates.forEach((d) => {
+          if (date.toDateString() === d.toDateString()) {
+            highlighted = true
+            return true
+          }
+        })
+      }
+
+      if (this.isDefined(this.highlighted.from) && this.isDefined(this.highlighted.to)) {
+        highlighted = date >= this.highlighted.from && date <= this.highlighted.to
+      }
+
+      if (typeof this.highlighted.days !== 'undefined' && this.highlighted.days.indexOf(date.getDay()) !== -1) {
+        highlighted = true
+      }
+
+      if (typeof this.highlighted.daysOfMonth !== 'undefined' && this.highlighted.daysOfMonth.indexOf(date.getDate()) !== -1) {
+        highlighted = true
+      }
+
+      if (typeof this.highlighted.customPredictor === 'function' && this.highlighted.customPredictor(date)) {
+        highlighted = true
+      }
+
+      return highlighted
+    },
+    dayClasses (day) {
+      return {
+        'selected': day.isSelected,
+        'disabled': day.isDisabled,
+        'highlighted': day.isHighlighted,
+        'today': day.isToday,
+        'weekend': day.isWeekend,
+        'sat': day.isSaturday,
+        'sun': day.isSunday,
+        'highlight-start': day.isHighlightStart,
+        'highlight-end': day.isHighlightEnd
+      }
+    },
+    /**
+     * Whether a day is highlighted and it is the first date
+     * in the highlighted range of dates
+     * @param {Date}
+     * @return {Boolean}
+     */
+    isHighlightStart (date) {
+      return this.isHighlightedDate(date) &&
+        (this.highlighted.from instanceof Date) &&
+        (this.highlighted.from.getFullYear() === date.getFullYear()) &&
+        (this.highlighted.from.getMonth() === date.getMonth()) &&
+        (this.highlighted.from.getDate() === date.getDate())
+    },
+    /**
+     * Whether a day is highlighted and it is the first date
+     * in the highlighted range of dates
+     * @param {Date}
+     * @return {Boolean}
+     */
+    isHighlightEnd (date) {
+      return this.isHighlightedDate(date) &&
+        (this.highlighted.to instanceof Date) &&
+        (this.highlighted.to.getFullYear() === date.getFullYear()) &&
+        (this.highlighted.to.getMonth() === date.getMonth()) &&
+        (this.highlighted.to.getDate() === date.getDate())
+    },
+    /**
+     * Helper
+     * @param  {mixed}  prop
+     * @return {Boolean}
+     */
+    isDefined (prop) {
+      return typeof prop !== 'undefined' && prop
+    }
+  }
+}
+</script>
