+++ test/unit/specs/PickerDay/initialData.spec.js
@@ -1,0 +1,26 @@
+import PickerDay from '@/components/PickerDay.vue'
+import {shallow} from '@vue/test-utils'
+import languages from '@/utils/DateLanguages'
+
+describe('PickerDay: DOM', () => {
+  let wrapper
+  beforeEach(() => {
+    wrapper = shallow(PickerDay, {
+      propsData: {
+        allowedToShowView: () => true,
+        translation: languages.translations.en,
+        pageDate: new Date(2018, 1, 1),
+        selectedDate: new Date(2018, 2, 24)
+      }
+    })
+  })
+
+  it('knows the selected date', () => {
+    const newDate = new Date(2016, 9, 15)
+    wrapper.setProps({
+      selectedDate: newDate
+    })
+    expect(wrapper.vm.isSelectedDate(newDate)).toEqual(true)
+    expect(wrapper.vm.isSelectedDate(new Date(2017, 1, 1))).toEqual(false)
+  })
+})
